backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.3.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
anthropic==0.5.0
python-dotenv==1.0.0
joblib==1.3.1
matplotlib==3.7.2
seaborn==0.12.2
python-multipart==0.0.6